requirements.txt
python-dotenv
fastapi
uvicorn[standard]
python-multipart
pdfplumber
requests
websockets
langchain
openai
youtube-transcript-api
chromadb
python-dotenv
